+++ core/src/index.js
@@ -42,3 +42,4 @@ export * from "./metrics/index.js";
 export * from "./continuation/index.js";
 export * from "./preflight/index.js";
 export * from "./compact/index.js";
+export * from "./response/index.js";